Move AuthRepositoryImpl.kt into the same file as AuthRepository.kt interface.
